Fix double free in test
the CUuidManager destructor got called twice.
Once from the test main binary and once from the libantibot.so

The libantibot.so already has all the polybob so it does not have to
call into the main binary anymore anyways. So the fix is to remove
polybob from the test main.

The same bug exists when running the server. To fix that I guess polybob
needs its own namespace.